pyproject.toml
[tool.poetry]
name = "flask-auth"
version = "0.1.0"
description = ""
authors = ["sergey <saalaus@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.0"
Flask-SQLAlchemy = "^2.5.1"
PyJWT = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.971"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
